Close popups on windowWillMiniaturize notification
We do this on windowDidMove/windowWillClose, but we don't on a 'will
miniaturize' notification. This can leave an application with an
orphan popup window, such as context menu which looks really weird.
I err on a safe side though - I don't close all popups - the application's
logic can be such that it has several windows and one of them gets
minimized (we can do this via QWidget's interface, for example) - would
be strange if all popups close. So I only close popups that have
the miniaturized window as a transient parent.

Task-number: QTBUG-77833
Change-Id: Ib655a27c0ce8661f9e7156e6035f7fffaff901b1
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>