Make sure QLatin1Literal fwd header is generated
This broke when QLatin1Literal got deprecated in change 45373c19243aea335897ba0f371a1dd53ae8f079

Both hunks in this patch are needed. The hunk in syncqt.pl removes the
QT_DEPRECATED_X(...) macro if it appears solely on a line in source code,
the second hunk inserts a newline after the QT_DEPRECATED_X(...) macro usage
to trigger that code path in the Perl script.

Before/after comparison of the headers generated in include/QtCore:

```
% diff  ~/old.txt ~/new.txt
105a106
> QLatin1Literal
```

Change-Id: I468dd2dd54bf115521ed82c6182236905556f568
Reviewed-by: Jörg Bornemann <joerg.bornemann@qt.io>